+++ CHANGES.rst
@@ -2,7 +2,7 @@
  CHANGES
 =========
 
-4.0.0 (unreleased)
+4.0.0 (2017-05-25)
 ==================
 
 - Add support for Python 3.4, 3.5, 3.6 and PyPy.